+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the esi-service project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### BREAKING CHANGES
+- Updated the minimum version of peerDependency axios to 0.21.1.
+
 ## [4.0.0] - 2020-04-19
 ### BREAKING CHANGES
 - Dropped Node.js 8 support.