+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the esi-service project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [6.0.0] - 2021-11-06
 ### BREAKING CHANGES
 - Dropped Node.js 10 support.
 
@@ -84,7 +85,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial codebase, copied from <https://github.com/Ionaru/eve-utils>.
 - Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/esi-service/compare/5.0.0...HEAD
+[Unreleased]: https://github.com/Ionaru/esi-service/compare/6.0.0...HEAD
+[6.0.0]: https://github.com/Ionaru/esi-service/compare/5.0.0...6.0.0
 [5.0.0]: https://github.com/Ionaru/esi-service/compare/4.0.0...5.0.0
 [4.0.0]: https://github.com/Ionaru/esi-service/compare/3.2.2...4.0.0
 [3.2.2]: https://github.com/Ionaru/esi-service/compare/3.2.1...3.2.2